% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_commit_push}
\alias{add_commit_push}
\title{Add, commit, and push a repository}
\usage{
add_commit_push(path = ".", commit_msg, use_github = TRUE)
}
\arguments{
\item{path}{A path to an existing local git repository.}

\item{commit_msg}{chracter string commit message.}

\item{use_github}{logical whether or not to push the commit to GitHub.}
}
\description{
Add, commit, and push a repository
}
