% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bad_link}
\alias{bad_link}
\title{A function for readability that handles checking of whether or not a link is functional}
\usage{
bad_link(link)
}
\arguments{
\item{link}{a URL}
}
\value{
boolean of whether or not the link is functional
}
\description{
A function for readability that handles checking of whether or not a link is functional
}
\author{
Ben Sabath
}
