% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_best_practices.R
\name{check_best_practices}
\alias{check_best_practices}
\title{Check if a package complies with the skeleton IQSS Best Practices}
\usage{
check_best_practices(path = ".", save_results = TRUE,
  report_results = TRUE, list_results = FALSE)
}
\arguments{
\item{path}{character string of the package path.}

\item{save_results}{logical whether or not to save the results of the
check in the package directory in a YAML file called
\code{.iqss.bestpractices.yml}.}

\item{report_results}{logical whether or not to report the results to the
console.}

\item{list_results}{logical whether or not to return the results as a list.}
}
\description{
Check if a package complies with the skeleton IQSS Best Practices
}
\examples{
\dontrun{
check_best_practices()
}

}
