% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_doc_links.R
\name{check_doc_links}
\alias{check_doc_links}
\title{Determine if URLs in a package's documentation or directory are valid}
\usage{
check_doc_links(path = ".", base_url)
}
\arguments{
\item{path}{A path pointing to an R package, or a directory containing R
documentation. The default is the current directory.}

\item{base_url}{a character string for a domain for which relative links are
subdomains. Used to resolve relative URL paths in documentation.}
}
\description{
Determine if URLs in a package's documentation or directory are valid
}
\examples{
check_doc_links()

}
