% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clean_links}
\alias{clean_links}
\title{Prepares relative links for checking}
\usage{
clean_links(links, base_url = "")
}
\arguments{
\item{links}{vector of URLs}

\item{base_url}{a character string of a domain for which relative links are
subdomains.}
}
\description{
Prepares relative links for checking
}
