% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_html_links}
\alias{get_html_links}
\title{Parses an html document and returns a vector of all links in the document}
\usage{
get_html_links(html_doc)
}
\arguments{
\item{html_doc}{A path to an html file}
}
\value{
Vector of urls
}
\description{
Parses an html document and returns a vector of all links in the document
}
\author{
Ben Sabath
}
