% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_doc_links.R
\name{get_rd_links}
\alias{get_rd_links}
\title{Extract links from Rd documentation files}
\usage{
get_rd_links(path, base_url)
}
\arguments{
\item{path}{character string path to an Rd R documentation file.}

\item{base_url}{a character string for a domain for which relative links are
subdomains. Used to resolve relative URL paths in documentation.}
}
\description{
Extract links from Rd documentation files
}
