% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_iqss_package.R
\name{init_iqss_package}
\alias{init_iqss_package}
\title{Initialize a new R package skeleton using IQSS best practices}
\usage{
init_iqss_package(path, author_name, description = getOption("devtools.desc"),
  use_rstudio = TRUE, github_auth_token = github_pat(), ...)
}
\arguments{
\item{path}{required character string of the location to create new package.
The last component of the path will be used as the package name.}

\item{author_name}{required character string listing the primary package
author's first and last name. Please also include an email address.
Should be in the format \code{"'Yoni Ben-Meshulam' <yoni@opower.com>"}}

\item{description}{list of DESCRIPTION values to override default values or
add additional values.}

\item{use_rstudio}{logical whether to create an Rstudio project file
\url{https://support.rstudio.com/hc/en-us/articles/200526207-Using-Projects}
(with \code{\link{use_rstudio}})?}

\item{github_auth_token}{Provide a personal access token (PAT) from
\url{https://github.com/settings/tokens}. Defaults to the \code{GITHUB_PAT}
environment variable. If \code{NULL} then only a local git repo is initialized,
not a remote GitHub repo.}

\item{...}{arguments to pass to methods.}
}
\description{
Initialize a new R package skeleton using IQSS best practices
}
\examples{
\dontrun{
init_iqss_package(path = 'mypkg',
                  author_name = "'Yoni Ben-Meshulam' <yoni@opower.com>"
}

}
