% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{init_test_suite}
\alias{init_test_suite}
\title{Initialise test suite skeleton}
\usage{
init_test_suite(pkg = ".", use_travis = TRUE, use_appveyor = TRUE, ...)
}
\arguments{
\item{pkg}{package description, can be path or package name. See
\code{\link{as.package}} for more information}

\item{use_travis}{logical whether or not to include a \code{.travis.yml} file
for using Travis CI.}

\item{use_appveyor}{logical whether or not to include a \code{appveyor.yml} file
for using Appveyor CI.}

\item{...}{arguments to pass to methods}
}
\description{
Initialise test suite skeleton
}
\examples{
\dontrun{
init_test_suite(pkg = 'mypackage')
}

}
