% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{multifile_link_parse}
\alias{multifile_link_parse}
\title{Parse documentation files in parallel}
\usage{
multifile_link_parse(paths)
}
\arguments{
\item{paths}{vector of file paths}
}
\description{
Parse documentation files in parallel
}
