% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{render_template}
\alias{render_template}
\title{Render IQSS template}
\source{
Modified from usethis.
}
\usage{
render_template(template, data = list())
}
\arguments{
\item{template}{character string identifying which template to render}

\item{data}{list with data to render in the template}
}
\description{
Render IQSS template
}
