% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{test_doc_links}
\alias{test_doc_links}
\title{Determine if all links in the package documentation are valid}
\usage{
test_doc_links(path = ".", base_url = "")
}
\arguments{
\item{path}{A path pointing to an R package, or a directory containing R documentation.
Default is the current directory.}

\item{base_url}{The url under which your documentation lives. If one is not provided, relative links in your documentation may be incorrectly listed as broken.}
}
\description{
Determine if all links in the package documentation are valid
}
