% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_doc_links.R
\name{try_links}
\alias{try_links}
\title{Contains code handling the determination of whether or not links are bad.}
\usage{
try_links(links)
}
\arguments{
\item{links}{a list of links processed by \code{clean_links}}
}
\description{
Contains code handling the determination of whether or not links are bad.
}
