% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_disqus.R
\name{use_disqus}
\alias{use_disqus}
\title{Add Disqus commenting integreation to a package's website landing page}
\usage{
use_disqus(pkg = ".", shortname, page_url, page_identifier)
}
\arguments{
\item{pkg}{package description, can be path or package name. See
\code{\link{as.package}} for more information}

\item{shortname}{character string with the website's Disqus short name}

\item{page_url}{character string with the website's URL}

\item{page_identifier}{character string with the website's Disqus page
identifier}
}
\description{
Add Disqus commenting integreation to a package's website landing page
}
\details{
Before using, visit
\url{https://disqus.com/profile/login/?next=/admin/create/} to register
your website with Disqus.
\code{use_disqus} assumes that your package documentation landing page
is generated from a file called index.Rmd, e.g. as created by
init_iqss_package or pkgdown.
}
