% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{uses_git}
\alias{uses_git}
\title{Check if a package uses git version control}
\source{
From devtools
}
\usage{
uses_git(path = ".")
}
\arguments{
\item{path}{character string of the package path.}
}
\description{
Check if a package uses git version control
}
