% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{uses_github}
\alias{uses_github}
\title{Check if a package uses github}
\source{
From devtools
}
\usage{
uses_github(path = ".")
}
\arguments{
\item{path}{character string of the package path.}
}
\description{
Check if a package uses github
}
