% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_iqss_package.R
\name{build_iqss_package}
\alias{build_iqss_package}
\title{Build an IQSS Best Practices compliant package, including dynamically
generating documentation and running tests}
\usage{
build_iqss_package(path = ".", rdocs = TRUE, rdocs_args = c("rd",
  "collate", "namespace"), readme = TRUE, website = TRUE, check = TRUE,
  git_commit = TRUE, commit_message = "IQSS package build",
  push_to_github = TRUE, ...)
}
\arguments{
\item{path}{character string of the package path.}

\item{rdocs}{logical whether or not to \code{roxegenize} the R documentation
and vignettes.}

\item{rdocs_args}{character vector of roclet names to use with package. This
defaults to \code{NULL}, which will use the \code{roclets} fields in the list
provided in the Roxygen \code{DESCRIPTION} field. If none are specified,
defaults to \code{c("collate", "namespace", "rd")}.}

\item{readme}{logical whether or not to render the READEM.Rmd.}

\item{website}{logical whether or not to render the package website.}

\item{check}{logical whether or not to use \code{\link{check_best_practices}}
to check that the package can be built and passes all tests. Also generates
an IQSS Report Card \url{https://github.com/IQSS/social_science_software_toolkit/blob/master/report_card/iqss_report_card_spec.md}.}

\item{git_commit}{logical whether or not to commit the build to git.}

\item{commit_message}{character string with a descriptive git commit message.
Only relevant if \code{git_commit = TRUE}.}

\item{push_to_github}{logical whether or not to push the commits to GitHub.
Only relevant if \code{git_commit = TRUE}.}

\item{...}{arguments to pass to methods.}
}
\description{
Build an IQSS Best Practices compliant package, including dynamically
generating documentation and running tests
}
\seealso{
\code{\link{document}}. \code{\link{render}},
\code{\link{build_site}}, \code{\link{check_best_practices}}
}
