% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_iqss_package.R
\name{build_iqss_package}
\alias{build_iqss_package}
\title{Build an IQSS Best Practices compliant package, including dynamically
generating documentation and running tests}
\usage{
build_iqss_package(path = ".")
}
\arguments{
\item{path}{character string of the package path}
}
\description{
Build an IQSS Best Practices compliant package, including dynamically
generating documentation and running tests
}
