% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_best_practices.R
\name{check_best_practices}
\alias{check_best_practices}
\title{Check if a package complies with the skeleton IQSS Best Practices for
Statistical Software Development}
\usage{
check_best_practices(path = ".", save_results = TRUE,
  report_results = TRUE, list_results = FALSE, run_cran_check = TRUE,
  calculate_coverage = TRUE)
}
\arguments{
\item{path}{character string of the package path.}

\item{save_results}{logical whether or not to save the results of the
check in the package directory in a YAML file called
\code{.iqss.bestpractices.yml}.}

\item{report_results}{logical whether or not to report the results to the
console.}

\item{list_results}{logical whether or not to return the results as a list.}

\item{run_cran_check}{logical whether or not to run CRAN check.}

\item{calculate_coverage}{logical whether or not to calculate test coverage.}
}
\value{
By default the function retuns a summary of the best practices check
to the console and a YAML formatted file called \code{.iqss_reportcard.yml} to the
package directory.It also updates the \code{.Rbuildignore} file so that it will
not conflict with the package build.
A list can also be returned with \code{list_results = TRUE}.
}
\description{
Check if a package complies with the skeleton IQSS Best Practices for
Statistical Software Development
}
\details{
Checks whether an R package is compliant with the \emph{minimal}
standards set out in the IQSS Best Practices for Software Development and
Sustainability.
}
\examples{
\dontrun{
check_best_practices()
}

}
\seealso{
The IQSS Report Card Spec can be found at:
\url{https://github.com/IQSS/social_science_software_toolkit/blob/master/report_card/iqss_report_card_spec.md}
}
