% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_doc_links.R
\name{clean_links}
\alias{clean_links}
\title{Prepares relative links for checking}
\usage{
clean_links(links, base_url = "")
}
\arguments{
\item{links}{vector of URLs}

\item{base_url}{a character string for a domain for which relative links are
subdomains. Used to resolve relative URL paths in documentation.}
}
\value{
A list of links, where each link is either a single link, or a vector
of a relative link attached to all possible base urls.
}
\description{
Prepares relative links for checking
}
