% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_doc_links.R
\name{get_file_list}
\alias{get_file_list}
\title{Gathers documentation files paths into a single vector}
\usage{
get_file_list(search_dirs)
}
\arguments{
\item{search_dirs}{vector of paths to directories to search for
documentation.}
}
\description{
Gathers documentation files paths into a single vector
}
