% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_doc_links.R
\name{get_html_links}
\alias{get_html_links}
\title{Parses an html document and returns a vector of all links in the document}
\usage{
get_html_links(html_doc, base_url = "")
}
\arguments{
\item{html_doc}{character string path to an html file}

\item{base_url}{a character string for a domain for which relative links are
subdomains. Used to resolve relative URL paths in documentation.}
}
\value{
a vector of urls
}
\description{
Parses an html document and returns a vector of all links in the document
}
\author{
Ben Sabath
}
