% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_md_links}
\alias{get_md_links}
\title{Parse markdown files to HTML and extract URLs}
\usage{
get_md_links(path, base_url = "")
}
\arguments{
\item{path}{character string path to a markdown file.}

\item{base_url}{a character string of a domain for which relative links are
subdomains.}
}
\description{
Parse markdown files to HTML and extract URLs
}
