% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_rd_links}
\alias{get_rd_links}
\title{Code to handle finding links in non-html documents.
Currently cannot handle relative links}
\usage{
get_rd_links(rd_doc)
}
\description{
Code to handle finding links in non-html documents.
Currently cannot handle relative links
}
