% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_rd_links}
\alias{get_rd_links}
\title{Extract links from Rd documentation files}
\usage{
get_rd_links(path)
}
\arguments{
\item{path}{character string path to an Rd R documentation file.}
}
\description{
Extract links from Rd documentation files
}
