% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_iqss_package.R
\name{init_iqss_package}
\alias{init_iqss_package}
\title{Initialize a new R package skeleton using IQSS best practices}
\usage{
init_iqss_package(path, description = getOption("devtools.desc"),
  use_rstudio = TRUE, use_pkgdown = TRUE, use_gpl3 = TRUE,
  use_tests = TRUE, use_git = TRUE, github_auth_token = github_pat(),
  github_protocol = "https", change_wd = TRUE, ...)
}
\arguments{
\item{path}{character string of the location to create new package.
The last component of the path will be used as the package name.}

\item{description}{list of DESCRIPTION values to override default values or
add additional values.}

\item{use_rstudio}{logical whether to create an Rstudio project file
\url{https://support.rstudio.com/hc/en-us/articles/200526207-Using-Projects}
(with \code{\link{use_rstudio}})?}

\item{use_pkgdown}{logical whether or not to initialize a pkgdown website
for documenting the package.}

\item{use_gpl3}{logical whether or not to include a GPL 3 license.}

\item{use_tests}{logical whether or not to initialize a test suite with the
testthat package and continuous integration with Travis CI (for Linux and
macOS) and Appveyor (for Windows).}

\item{use_git}{logical whether or not to use git version control.}

\item{github_auth_token}{Provide a personal access token (PAT) from
\url{https://github.com/settings/tokens}. Defaults to the \code{GITHUB_PAT}
environment variable. If \code{NULL} then only a local git repo is initialized,
not a remote GitHub repo.}

\item{github_protocol}{character string specifying the transfer protocol for pushing
the GitHub remote repository , either \code{"ssh"} (the default) or \code{"https"}.
Only relevant if \code{github_auth_token} is provided.}

\item{change_wd}{logical whether to change to make the package the working
directory upon completion.}

\item{...}{arguments to pass to methods.}
}
\description{
Initialize a new R package skeleton using IQSS best practices
}
\examples{
\dontrun{
# Initialize new package called "mypkg" in the current working directory
init_iqss_package(path = 'mypkg',
                  description = list("Title" = "A Test IQSS Package",
                                     "Description" =
                                     "This is a longer description of the package."))
}

}
\seealso{
\code{\link{create}} \code{\link{github_pat}}
\code{\link{use_news_md}}
\code{\link{use_gpl3_license}} \code{\link{use_github}}
\code{\link{use_git}} \code{\link{use_testthat}} \code{\link{use_travis}}
\code{\link{use_appveyor}} \code{\link{use_package_doc}},
\code{\link{init_site}}
}
