% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_doc_links.R
\name{parse_file}
\alias{parse_file}
\title{Extract URL links from documentation files}
\usage{
parse_file(path, base_url)
}
\arguments{
\item{path}{character string of file path to extract links from}

\item{base_url}{a character string for a domain for which relative links are
subdomains. Used to resolve relative URL paths in documentation.}
}
\description{
Extract URL links from documentation files
}
\details{
Currently extracts links for HTML, Rd, and markdown files. Allows
for future addition of file types if needed
}
