% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{test_doc_links}
\alias{test_doc_links}
\title{Determine if URLs in a package's documentation or directory are valid}
\usage{
test_doc_links(path = ".", base_url = "")
}
\arguments{
\item{path}{A path pointing to an R package, or a directory containing R
documentation. The default is the current directory.}

\item{base_url}{a character string of a domain for which relative links are
subdomains.}
}
\description{
Determine if URLs in a package's documentation or directory are valid
}
